# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Documentation of our awesome API.",
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/v1",
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "description": "Need to send new password",
        "tags": [
          "Auth"
        ],
        "operationId": "ForgotPassword",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "x-go-name": "Code"
                },
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                },
                "new_password": {
                  "type": "string",
                  "x-go-name": "New_password"
                },
                "new_password_second": {
                  "type": "string",
                  "x-go-name": "New_password_second"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
    "/auth/forgot-password-code": {
      "post": {
        "description": "Need to send new password",
        "tags": [
          "Auth"
        ],
        "operationId": "ForgotPasswordCode",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
    "/refresh-token": {
      "get": {
        "description": "Send request within refresh-token to get new access-token",
        "tags": [
          "Auth"
        ],
        "operationId": "RefreshToken",
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "401": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
    "/set-password": {
      "post": {
        "description": "Need to send old pass/code/new_password",
        "tags": [
          "Auth"
        ],
        "operationId": "SetNewPassword",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "new_password": {
                  "type": "string",
                  "x-go-name": "New_password"
                },
                "new_password_second": {
                  "type": "string",
                  "x-go-name": "New_password_second"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
    "/user/{user_id}": {
      "get": {
        "description": "You receive an list of users",
        "tags": [
          "User"
        ],
        "operationId": "GetUserById",
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      },
      "put": {
        "description": "You receive an list of users",
        "tags": [
          "User"
        ],
        "operationId": "UpdateUserById",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "addres": {
                  "type": "string",
                  "x-go-name": "Address"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "username": {
                  "type": "string",
                  "x-go-name": "Username"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
    "/verify/mail": {
      "get": {
        "description": "Recieve mail with a:href to verify email addres",
        "tags": [
          "Auth"
        ],
        "operationId": "VerifyMail",
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "401": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "GenericResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "x-go-name": "Data"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "boolean",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "_/home/dzemil/Projects/golang/risc_monolith/server/utils"
    },
...
    "showUsers": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id"
        },
        "isverified": {
          "type": "boolean",
          "x-go-name": "Isverified"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "role": {
          "type": "string",
          "x-go-name": "Role"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "_/home/dzemil/Projects/golang/risc_monolith/server/utils"
    },
...
  },
  "responses": {
...
    "genericResponse": {
      "description": "This text will appear as description of your response body.",
      "schema": {
        "$ref": "#/definitions/GenericResponse"
      }
    },
...
    "userResponse": {
      "description": "User with basic info",
      "schema": {
        "type": "object",
        "properties": {
          "User": {
            "$ref": "#/definitions/showUsers"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}